server/qserver/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
blis==0.7.4
bs4==0.0.1
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cymem==2.0.5
Django==3.1.7
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-tracking==1.5.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
metadata-parser==0.10.5
murmurhash==1.0.5
mysql-connector-python==8.0.25
numpy==1.21.1
opengraph-py3==0.71
packaging==21.0
pathy==0.6.0
Pillow==8.3.1
pio==0.0.3
preshed==3.0.5
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-magic-bin==0.4.14
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
smart-open==5.1.0
soupsieve==2.2.1
sqlparse==0.4.1
srsly==2.4.1
thinc==8.0.8
topicaxis-opengraph==0.5
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
